One-click stages, a silhouette-only hull field, scripted retopology
stage1.py takes a mesh to images: 14 orthographic views with anti-aliased
masks, four-light Lambertian renders of each, and the normals solved from
them. It is the only step that reads the mesh.

stage2.py takes those images to a mesh: hull field, per-view anchored depth
(three views at a time), continuous-field fusion, retopology, and -- only if
a ground-truth mesh is named -- scoring and showcase renders. Each step is
skipped when its output exists, so a long run resumes.

tools/hull_field.py builds the hull without any metadata: the grid comes from
the axis-aligned silhouettes' extents, H is evaluated exactly only in a band
around the surface (it is 1-Lipschitz, so a stride-4 coarse pass bounds
everything else), and each view's hull depth is found by sphere tracing H.
On Lucy the hull depth is in front of the true surface at 98.8% to 100% of
pixels per view; the rest are rim pixels, where a 0.3-voxel radial slack
becomes a few voxels along a grazing ray.

tools/retopo.py: voxel remesh to a clean manifold, keep the largest shell,
QuadriFlow to a quad base, then subdivide and shrink-wrap onto the full
surface. Two things made QuadriFlow refuse silently until fixed: internal
voids closed into inward-facing shells, and its check rejecting any edge
shorter than 1e-4 object units, which a one-unit-tall model is full of.

tools/prep_mesh.py normalises every Stanford PLY encoding (ASCII, big-endian,
extra vertex and face properties) to little-endian triangles with +Z up.

The sweep's cost volume is float16 so three views fit in memory at once.

+++ tools/depth_mv.py
@@ -47,9 +47,9 @@ def refine_parabola(cost, k, offsets):
     K = cost.shape[0]
     km = np.clip(k - 1, 0, K - 1)
     kp = np.clip(k + 1, 0, K - 1)
-    c0 = np.take_along_axis(cost, k[None], 0)[0]
-    cm = np.take_along_axis(cost, km[None], 0)[0]
-    cp = np.take_along_axis(cost, kp[None], 0)[0]
+    c0 = np.take_along_axis(cost, k[None], 0)[0].astype(np.float32)
+    cm = np.take_along_axis(cost, km[None], 0)[0].astype(np.float32)
+    cp = np.take_along_axis(cost, kp[None], 0)[0].astype(np.float32)
     den = cm - 2 * c0 + cp
     ok = np.isfinite(cm) & np.isfinite(cp) & (den > 1e-9) & (k > 0) & (k < K - 1)
     delta = np.where(ok, 0.5 * (cm - cp) / np.where(ok, den, 1), 0.0)
@@ -124,7 +124,7 @@ def process(views_dir, hull_dir, normals_dir, view, meta, args, nB, hitB):
     cost = ns.sweep(meta, view, srcs, nA, nB, hitB, relief, hull, hit,
                     offsets, args.win)
     k = np.argmin(cost, axis=0)
-    best = np.take_along_axis(cost, k[None], 0)[0]
+    best = np.take_along_axis(cost, k[None], 0)[0].astype(np.float32)
     off = refine_parabola(cost, k, offsets)
     del cost
     z_st = relief + off
